1.0

+++ app/Http/Controllers/DebtorController.php
@@ -127,6 +127,8 @@ class DebtorController extends Controller
     public function destroy($id)
     {
         $debtor = Debtor::find($id);
-        $debtor->delete();
+        if ($debtor->account != 0) {
+            $debtor->delete();
+        }
     }
 }
